add tests for `commands/typedoc`
TODO: ensure dirnameStub and mkdirpStub are being called. For some
reason, our `Typedoc` class is not being called and then calling the
`super` method for `generateJson` and `generateDocs`. Rather, it seems
that the `TypedocApp` class method stubs are being called directly.
Until this is resolved, there are two assertions commented out.