finish test for `tasks/util/wrapAsyncTask`
change mocking strategy to provide only basic config and mock needed to
run the `sync` task as opposed to creating a simple task in the test
suite. because of the way Grunt works (loading config by string name),
when mocking a task in the suite there are errors as the function for
the task is never actually called.